Move sample instances for testing in a separate object.
