+++ src/test/scala/mqtt/broker/TestBrokerState.scala
@@ -1,41 +1,10 @@
 package mqtt.broker
 
-import java.util.Calendar
-
+import mqtt.broker.SampleInstances._
 import org.scalatest.FunSuite
 
-import scala.concurrent.duration.Duration
-
 class TestBrokerState extends FunSuite {
-  val sample_session_1 = Session(
-    socket = Option.empty,
-    keepAlive = Duration(0, "millis"),
-    lastContact = Calendar.getInstance().getTime,
-    subscriptions = Map(),
-    notYetAcknowledged = Map(),
-    receivedButNotYetAcknowledged = Map(),
-    pendingTransmission = Seq(),
-    persistent = false
-  )
   
-  val sample_session_2 = Session(
-    socket = Option.empty,
-    keepAlive = Duration(10, "minutes"),
-    lastContact = Calendar.getInstance().getTime,
-    subscriptions = Map(),
-    notYetAcknowledged = Map(),
-    receivedButNotYetAcknowledged = Map(),
-    pendingTransmission = Seq(),
-    persistent = false
-  )
-  
-  val sample_id_1 = "123"
-  val sample_id_2 = "456"
-  
-  val sample_socket_1 = Socket(1, Option.empty)
-  val sample_socket_2 = Socket(2, Option.empty)
-  
-  val bs0 = BrokerState(Map(), Map(), Map())
   
   //TODO addClosingChannel test
   //TODO add tests for yet untested methods
